requirements.txt
fastmcp
httpx
aiofiles
python-dotenv
